Revert unnecessary changes from 22019 (#24802)
At the time, https://github.com/pandas-dev/pandas/pull/22019 seemed to
break API in a few places. Since then, those API breaking changes have
disappeared, so the changes to the tests and the release note are no
longer necessary.

Closes https://github.com/pandas-dev/pandas/issues/24789

+++ pandas/tests/frame/test_axis_select_reindex.py
@@ -717,7 +717,7 @@ class TestDataFrameSelectReindex(TestData):
 
         result = df1 - df1.mean()
         expected = df2 - df2.mean()
-        assert_frame_equal(result.astype('f8'), expected)
+        assert_frame_equal(result, expected)
 
     def test_align_multiindex(self):
         # GH 10665
